upd: Organiza as tools do fluxo de codificação em um mesmo modulo
